# Conflicts:
#	lect-presenter-api/src/main/java/org/lecturestudio/presenter/api/presenter/SlidesPresenter.java
#	lect-presenter-swing/src/main/java/org/lecturestudio/presenter/swing/view/SwingSlidesView.java
#	lect-presenter-swing/src/main/resources/views/main-slides/main-slides_de_DE.properties
#	lect-presenter-swing/src/main/resources/views/main-slides/main-slides_en_US.properties

+++ lect-presenter-fx/src/main/java/org/lecturestudio/presenter/javafx/view/FxSlidesView.java
@@ -421,7 +421,7 @@ public class FxSlidesView extends VBox implements SlidesView {
 	}
 
 	@Override
-	public void setOnStopPeerConnection(ConsumerAction<BigInteger> action) {
+	public void setOnStopPeerConnection(ConsumerAction<Long> action) {
 
 	}
